+++ lib/src/controls/inputs/slider.dart
@@ -259,7 +259,6 @@ class _SliderState extends State<Slider> {
 
   void _showLabelOverlay() {
     final sliderState = materialSliderKey.currentState as dynamic;
-    sliderState.showValueIndicator();
     (sliderState.overlayController as AnimationController).forward();
     (sliderState.valueIndicatorController as AnimationController).forward();
   }
@@ -298,7 +297,7 @@ class _SliderState extends State<Slider> {
           ),
           builder: (context, innerFactor, child) => m.SliderTheme(
             data: m.SliderThemeData(
-              showValueIndicator: m.ShowValueIndicator.always,
+              showValueIndicator: m.ShowValueIndicator.onDrag,
               thumbColor: style.thumbColor?.resolve(states),
               overlayShape: const m.RoundSliderOverlayShape(overlayRadius: 0),
               valueIndicatorTextStyle: TextStyle(